refactor: extract shared claim text utilities into crux/lib/claim-text-utils.ts
Consolidate duplicated markup stripping, entity reference detection, and
slug-to-display-name helpers from 6 files into a single shared module.
Removes ~270 lines of duplicated code that had already started to drift.

Files updated:
- extraction-quality-gate.ts: removed local copies, imports from shared module
- fix-all.ts: uses shared stripMarkup and escapeRegex
- fix-quality.ts: uses shared stripMarkup via thin wrapper
- quality-report.ts: uses shared hasMarkup
- validate-quality/checks.ts: uses shared containsEntityReference,
  isTautologicalDefinition, escapeRegex, slugToDisplayName
- extract.ts: uses shared slugToDisplayName
